Merge Development into Master
Merge development: add REST API endpoints, improve Selenium tests, and refactor configuration handling

+++ scripts/ci-cd.sh
@@ -5,11 +5,12 @@
 
 # The task of this script is to test the current commit in the testing branch and, if successful, to move it to the main branch.
 # git pull origin testing -> selenium tests (integration tests) -> git merge into master -> git push origin master -> cd production -> git pull origin master (update in production)
-
+(
+  flock -n 200 || { echo "Another instance is already running. Exiting."; exit 1; }
 # setup directories
 repo_dir="/home/git/repositories/dienstplan-apotheke-testing"
 hostnameInstallTest="https://docker.martin-mandelkow.de"
-export JAVA_HOME=/usr/lib/jvm/jre-11-openjdk/
+export JAVA_HOME=/usr/lib/jvm/default-java
 ENVIRONMENT=testing # set ENVIRONMENT for the Dockerfile to testing
 
 # Determine the script's directory
@@ -28,10 +29,10 @@ source ~/.bash_profile
 
 rm -rf "$repo_dir"
 mkdir -p "$repo_dir"
-cd "$repo_dir" || exit 
+cd "$repo_dir" || exit
 # clone repository and checkout testing branch
 git clone git@github.com:MaMaKow/dienstplan-apotheke.git
-cd dienstplan-apotheke || exit 
+cd dienstplan-apotheke || exit
 git checkout testing
 git fetch --all
 # Fetch latest changes from remote
@@ -45,17 +46,49 @@ if ! git merge --no-ff origin/master; then
 fi
 echo "Merge with master completed successfully. Proceeding with the script."
 
+# Get the 3rdparty classes via composer:
+composer dump-autoload
+composer install
+#composer install --no-dev --optimize-autoloader
+
 # Check for differences between testing and master branches
 diff_output=$(git diff origin/master..origin/testing)
 if [ -z "$diff_output" ]; then
     echo "No differences between testing and master branches. Exiting."
     exit 0
 fi
+# Check if docker-compose file is existent
+if [ ! -f "$repo_dir/dienstplan-apotheke/docker-compose.yml" ]; then
+    echo "docker-compose.yml not found. Exiting."
+    exit 1
+fi
+
+
+echo HERE1
 # Get valid certificate files
 # PDR requires a connection via HTTPS. Certificates are therefore required.
-# These must be provided in the same path as this script. They are copied from here into the docker container. 
+# These must be provided in the same path as this script. They are copied from here into the docker container.
 cp $script_dir/fullchain.pem $repo_dir/dienstplan-apotheke/upload/fullchain.pem
 cp $script_dir/privkey.pem $repo_dir/dienstplan-apotheke/upload/privkey.pem
+echo HERE2
+
+
+# Install dependencies with composer
+# First get composer:
+EXPECTED_HASH=$(php -r "copy('https://composer.github.io/installer.sig', 'php://stdout');")
+php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
+ACTUAL_HASH=$(php -r "echo hash_file('SHA384', 'composer-setup.php');")
+
+if [ "$EXPECTED_HASH" != "$ACTUAL_HASH" ]; then
+  echo "ERROR: Invalid installer hash"
+  rm composer-setup.php
+  exit 1
+fi
+
+php composer-setup.php
+rm composer-setup.php
+# Second run composer to install the dependencies:
+php composer.phar install
 
 # configure environment for docker container
 # expose a random port and chosse random user names and passphrases
@@ -72,6 +105,7 @@ echo MYSQL_DATABASE=$random_db_name >> .env
 echo MYSQL_USER=$random_user_name >>.env
 echo MYSQL_PASSWORD=$random_user_passphrase >> .env
 urlInstallTest=$hostnameInstallTest:$random_secure_web_port/apotheke
+echo HERE3
 
 # The databaseHostname is "db". This is the hostname of the mysql container as defined in the docker-compose.yml
 # TODO: Use databaseUserName and passphrase from random instead of root
@@ -91,8 +125,11 @@ databasePassword=$random_root_passphrase
 databaseName=$random_db_name
 databasePort=3306
 EOF
+echo HERE4
 #cat $repo_dir/dienstplan-apotheke/tests/selenium/Configuration.properties
+echo HERE5
 bash $repo_dir/dienstplan-apotheke/scripts/restart_docker_container.sh
+echo HERE6
 #docker-compose build --no-cache
 docker-compose -f "$repo_dir/dienstplan-apotheke/docker-compose.yml" down --volumes
 docker-compose -f "$repo_dir/dienstplan-apotheke/docker-compose.yml" build
@@ -100,7 +137,7 @@ docker-compose -f "$repo_dir/dienstplan-apotheke/docker-compose.yml" up -d
 # Function to check if all containers are running
 check_containers() {
   # Get the status of all containers
-  STATUS=$(docker-compose ps -q | xargs docker inspect -f '{{ .State.Running }}' 2>/dev/null)
+  STATUS=$(docker-compose -f "$repo_dir/dienstplan-apotheke/docker-compose.yml" ps -q | xargs docker inspect -f '{{ .State.Running }}' 2>/dev/null)
 
   # Check if any container is not running
   if [[ "$STATUS" == *"false"* ]] || [[ -z "$STATUS" ]]; then
@@ -112,19 +149,28 @@ check_containers() {
 
 # Wait for all containers to be running
 echo "Waiting for containers to be up."
+number_of_times_containers_checked=0;
 while ! check_containers; do
     echo -n "."
+    (( number_of_times_containers_checked++ ))
+    if [ $number_of_times_containers_checked -gt 30 ]; then
+	echo "Taking to long to wait for container. Exiting."
+        exit 1;
+    fi
     sleep 1  # Wait for 1 second before checking again
 done
 #sleep 30
 
 # run selenium tests
 # assuming the selenium tests are written to exit with a non-zero status on failure
-cd "$repo_dir"/dienstplan-apotheke/tests/selenium/ || exit 
+cd "$repo_dir"/dienstplan-apotheke/tests/selenium/ || exit
+
+echo Test connection using curl:
+curl -vvv https://$urlInstallTest/dienstplan-test/
+
+
 /usr/bin/mvn test | tee ./mvn.log
 echo -e "\a" # Bell sound!
-# cleanup the docker container
-docker-compose -f "$repo_dir/dienstplan-apotheke/docker-compose.yml" down --volumes
 
 test_outcome=$(cat test-result)
 if [ "$test_outcome" == "FAILED" ]; then
@@ -134,8 +180,15 @@ elif [ "$test_outcome" == "SUCCESS" ]; then
     echo "Selenium tests succeeded."
 else
     echo "Unexpected result in test-result file: $test_outcome"
+    # cleanup the docker container in case of failure:
+    docker-compose -f "$repo_dir/dienstplan-apotheke/docker-compose.yml" down --volumes
     exit 1
 fi
+docker-compose -f "$repo_dir/dienstplan-apotheke/docker-compose.yml" ps -q db
+docker exec "$test_db_container" mysqldump -u root -p"$random_root_passphrase" "$random_db_name" > /tmp/db_dump.sql
+# cleanup the docker container also in case of success:
+docker-compose -f "$repo_dir/dienstplan-apotheke/docker-compose.yml" down --volumes
+bash "$repo_dir/dienstplan-apotheke/scripts/start-live-test-container.sh"
 
 # if tests pass, merge into master and push
 # Check the current path
@@ -156,5 +209,5 @@ git push https://$GIT_TOKEN@github.com/MaMaKow/dienstplan-apotheke.git testing:m
 
 echo "CI/CD pipeline executed successfully."
 # finally delete the testing directory
-#rm -rf "$repo_dir"
-
+rm -rf "$repo_dir"
+) 200>/home/git/scripts/ci-cd.lock
